super-prodzekt-nx/libs/handlownia/feature/feature-product-displayer/src/lib/components/displayer/product-details/product-details.component.ts
import {ChangeDetectionStrategy, Component, EventEmitter, Input, Output} from '@angular/core';
import {Product} from "../../../models/product";

@Component({
  selector: 'super-prodzekt-nx-product-details',
  templateUrl: './product-details.component.html',
  styleUrls: ['./product-details.component.scss'],
  changeDetection: ChangeDetectionStrategy.OnPush,
})
export class ProductDetailsComponent {

  @Input() product: Product = {
    id: -1,
    header: 'header',
    description: 'description',
    imageUrl: 'imageUrl',
    isPromoted: false,
  };

  @Output() productClicked = new EventEmitter<void>();

  openProductDetails() {
    this.productClicked.emit();
  }
}
